chore(test): migrate external-imports
Also a `await $('#item-4').waitForExist();` in `test/wdio/key-reorder/cmp.test.tsx`, which was contributed in #5472, as the test seemed to be flaky on my machine. The waitForExist help to remove this flakiness.
