--- test/karma/test-app/external-imports/index.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf8">
-<script src="./build/testapp.esm.js" type="module"></script>
-<script src="./build/testapp.js" nomodule></script>
-
-<external-import-a></external-import-a>
-<external-import-b></external-import-b>
-<external-import-c></external-import-c>
